Migrate to pyproject.toml
