+++ .gitignore
@@ -105,3 +105,6 @@ venv.bak/
 
 # VS Code
 .vscode/
+
+# Version file
+python/src/workshop/_version.py
