+++ python/MANIFEST.in
@@ -1,2 +1,1 @@
-include versioneer.py
 include odin_workshop/_version.py
