+++ python/pyproject.toml
@@ -1,0 +1,31 @@
+[build-system]
+requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "workshop"
+classifiers = [
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12"
+]
+dynamic = ["version"]
+description = "ODIN Workshop"
+license.file = "LICENSE"
+readme = "README.md"
+authors = [
+    {name = "Tim Nicholls", email = "tim.nicholls@stfc.ac.uk"}
+]
+dependencies = [
+    "odin_control @ git+https://git@github.com/odin-detector/odin-control.git@1.6.0",
+    "tornado>=4.3",
+    "future"
+]
+requires-python = ">=3.8"
+
+[tool.setuptools_scm]
+version_file = "src/workshop/_version.py"
+root = ".."
