--- python/setup.cfg
@@ -1,23 +1,0 @@
-[nosetests]
-verbosity=2
-nocapture=1
-detailed-errors=1
-with-coverage=1
-cover-package=odin_workshop
-cover-erase=1
-#debug=nose.loader
-#pdb=1
-#pdb-failures=1
-
-[flake8]
-max-line-length = 100
-
-[pycodestyle]
-max-line-length = 100
-
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = odin_workshop/_version.py
-versionfile_build = odin_workshop/_version.py
-tag_prefix=
