+++ python/src/workshop/adapter.py
@@ -1,0 +1,118 @@
+"""Demo adapter for ODIN control workshop
+
+This class implements a simple adapter used for demonstration purposes in a
+
+Tim Nicholls, STFC Application Engineering
+"""
+import logging
+import sys
+
+from tornado.escape import json_decode
+
+from odin.adapters.adapter import ApiAdapter, ApiAdapterResponse, request_types, response_types
+from odin.adapters.parameter_tree import ParameterTreeError
+
+from workshop.controller import WorkshopController, WorkshopError
+
+
+class WorkshopAdapter(ApiAdapter):
+    """System info adapter class for the ODIN server.
+
+    This adapter provides ODIN clients with information about the server and the system that it is
+    running on.
+    """
+
+    def __init__(self, **kwargs):
+        """Initialize the WorkshopAdapter object.
+
+        This constructor initializes the WorkshopAdapter object.
+
+        :param kwargs: keyword arguments specifying options
+        """
+        # Intialise superclass
+        super(WorkshopAdapter, self).__init__(**kwargs)
+
+        # Parse options
+        background_task_enable = bool(self.options.get('background_task_enable', False))
+        background_task_interval = float(self.options.get('background_task_interval', 1.0))
+
+        self.controller = WorkshopController(background_task_enable, background_task_interval)
+
+        logging.debug('WorkshopAdapter loaded')
+
+    @response_types('application/json', default='application/json')
+    def get(self, path, request):
+        """Handle an HTTP GET request.
+
+        This method handles an HTTP GET request, returning a JSON response.
+
+        :param path: URI path of request
+        :param request: HTTP request object
+        :return: an ApiAdapterResponse object containing the appropriate response
+        """
+        try:
+            response = self.controller.get(path)
+            status_code = 200
+        except ParameterTreeError as e:
+            response = {'error': str(e)}
+            status_code = 400
+
+        content_type = 'application/json'
+
+        return ApiAdapterResponse(response, content_type=content_type,
+                                  status_code=status_code)
+
+    @request_types('application/json')
+    @response_types('application/json', default='application/json')
+    def put(self, path, request):
+        """Handle an HTTP PUT request.
+
+        This method handles an HTTP PUT request, returning a JSON response.
+
+        :param path: URI path of request
+        :param request: HTTP request object
+        :return: an ApiAdapterResponse object containing the appropriate response
+        """
+
+        content_type = 'application/json'
+
+        try:
+            data = json_decode(request.body)
+            self.controller.set(path, data)
+            response = self.controller.get(path)
+            status_code = 200
+        except WorkshopError as e:
+            response = {'error': str(e)}
+            status_code = 400
+        except (TypeError, ValueError) as e:
+            response = {'error': 'Failed to decode PUT request body: {}'.format(str(e))}
+            status_code = 400
+
+        logging.debug(response)
+
+        return ApiAdapterResponse(response, content_type=content_type,
+                                  status_code=status_code)
+
+    def delete(self, path, request):
+        """Handle an HTTP DELETE request.
+
+        This method handles an HTTP DELETE request, returning a JSON response.
+
+        :param path: URI path of request
+        :param request: HTTP request object
+        :return: an ApiAdapterResponse object containing the appropriate response
+        """
+        response = 'WorkshopAdapter: DELETE on path {}'.format(path)
+        status_code = 200
+
+        logging.debug(response)
+
+        return ApiAdapterResponse(response, status_code=status_code)
+
+    def cleanup(self):
+        """Clean up adapter state at shutdown.
+
+        This method cleans up the adapter state when called by the server at e.g. shutdown.
+        It simplied calls the cleanup function of the controller instance.
+        """
+        self.controller.cleanup()
