+++ python/src/workshop/controller.py
@@ -1,125 +1,15 @@
-"""Demo adapter for ODIN control workshop
-
-This class implements a simple adapter used for demonstration purposes in a
-
-Tim Nicholls, STFC Application Engineering
-"""
-import logging
 import tornado
 import time
 import sys
+import logging
 from concurrent import futures
 
 from tornado.ioloop import IOLoop, PeriodicCallback
 from tornado.concurrent import run_on_executor
-from tornado.escape import json_decode
 
-from odin.adapters.adapter import ApiAdapter, ApiAdapterResponse, request_types, response_types
 from odin.adapters.parameter_tree import ParameterTree, ParameterTreeError
-from odin._version import get_versions
 
-
-class WorkshopAdapter(ApiAdapter):
-    """System info adapter class for the ODIN server.
-
-    This adapter provides ODIN clients with information about the server and the system that it is
-    running on.
-    """
-
-    def __init__(self, **kwargs):
-        """Initialize the WorkshopAdapter object.
-
-        This constructor initializes the WorkshopAdapter object.
-
-        :param kwargs: keyword arguments specifying options
-        """
-        # Intialise superclass
-        super(WorkshopAdapter, self).__init__(**kwargs)
-
-        # Parse options
-        background_task_enable = bool(self.options.get('background_task_enable', False))
-        background_task_interval = float(self.options.get('background_task_interval', 1.0))
-
-        self.workshop = Workshop(background_task_enable, background_task_interval)
-
-        logging.debug('WorkshopAdapter loaded')
-
-    @response_types('application/json', default='application/json')
-    def get(self, path, request):
-        """Handle an HTTP GET request.
-
-        This method handles an HTTP GET request, returning a JSON response.
-
-        :param path: URI path of request
-        :param request: HTTP request object
-        :return: an ApiAdapterResponse object containing the appropriate response
-        """
-        try:
-            response = self.workshop.get(path)
-            status_code = 200
-        except ParameterTreeError as e:
-            response = {'error': str(e)}
-            status_code = 400
-
-        content_type = 'application/json'
-
-        return ApiAdapterResponse(response, content_type=content_type,
-                                  status_code=status_code)
-
-    @request_types('application/json')
-    @response_types('application/json', default='application/json')
-    def put(self, path, request):
-        """Handle an HTTP PUT request.
-
-        This method handles an HTTP PUT request, returning a JSON response.
-
-        :param path: URI path of request
-        :param request: HTTP request object
-        :return: an ApiAdapterResponse object containing the appropriate response
-        """
-
-        content_type = 'application/json'
-
-        try:
-            data = json_decode(request.body)
-            self.workshop.set(path, data)
-            response = self.workshop.get(path)
-            status_code = 200
-        except WorkshopError as e:
-            response = {'error': str(e)}
-            status_code = 400
-        except (TypeError, ValueError) as e:
-            response = {'error': 'Failed to decode PUT request body: {}'.format(str(e))}
-            status_code = 400
-
-        logging.debug(response)
-
-        return ApiAdapterResponse(response, content_type=content_type,
-                                  status_code=status_code)
-
-    def delete(self, path, request):
-        """Handle an HTTP DELETE request.
-
-        This method handles an HTTP DELETE request, returning a JSON response.
-
-        :param path: URI path of request
-        :param request: HTTP request object
-        :return: an ApiAdapterResponse object containing the appropriate response
-        """
-        response = 'WorkshopAdapter: DELETE on path {}'.format(path)
-        status_code = 200
-
-        logging.debug(response)
-
-        return ApiAdapterResponse(response, status_code=status_code)
-
-    def cleanup(self):
-        """Clean up adapter state at shutdown.
-
-        This method cleans up the adapter state when called by the server at e.g. shutdown.
-        It simplied calls the cleanup function of the workshop instance.
-        """
-        self.workshop.cleanup()
+from workshop._version import __version__
 
 class WorkshopError(Exception):
     """Simple exception class to wrap lower-level exceptions."""
@@ -127,16 +17,16 @@ class WorkshopError(Exception):
     pass
 
 
-class Workshop():
-    """Workshop - class that extracts and stores information about system-level parameters."""
+class WorkshopController():
+    """WorkshopController - class that extracts and stores information about system-level parameters."""
 
     # Thread executor used for background tasks
     executor = futures.ThreadPoolExecutor(max_workers=1)
 
     def __init__(self, background_task_enable, background_task_interval):
-        """Initialise the Workshop object.
+        """Initialise the WorkshopController object.
 
-        This constructor initlialises the Workshop object, building a parameter tree and
+        This constructor initlialises the WorkshopController object, building a parameter tree and
         launching a background task if enabled
         """
         # Save arguments
@@ -147,7 +37,6 @@ class Workshop():
         self.init_time = time.time()
 
         # Get package version information
-        version_info = get_versions()
 
         # Set the background task counters to zero
         self.background_ioloop_counter = 0
@@ -163,7 +52,7 @@ class Workshop():
 
         # Store all information in a parameter tree
         self.param_tree = ParameterTree({
-            'odin_version': version_info['version'],
+            'odin_version': __version__,
             'tornado_version': tornado.version,
             'server_uptime': (self.get_server_uptime, None),
             'background_task': bg_task 
@@ -204,7 +93,7 @@ class Workshop():
             raise WorkshopError(e)
 
     def cleanup(self):
-        """Clean up the Workshop instance.
+        """Clean up the WorkshopController instance.
 
         This method stops the background tasks, allowing the adapter state to be cleaned up
         correctly.
